requirements.txt
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
backports.tarfile==1.2.0
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cleo==2.1.0
crashtest==0.4.1
cryptography==45.0.3
cssselect2==0.8.0
distlib==0.3.9
Django>=4.2,<5.0
djangorestframework>=3.14.0
drf-yasg>=1.21.7
dulwich==0.22.8
exceptiongroup==1.3.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
fonttools==4.58.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
lxml==5.4.0
more-itertools==10.7.0
msgpack==1.1.0
oscrypto==1.3.0
packaging==25.0
pbs-installer==2025.5.17
Pillow>=10.2.0
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pypdf==5.6.0
pyphen==0.17.2
pyproject_hooks==1.2.0
python-bidi==0.6.6
python-dotenv>=1.0.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
tomlkit==0.13.2
trove-classifiers==2025.5.9.12
typing_extensions==4.13.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==5.0.0
urllib3==2.4.0
virtualenv==20.31.2
weasyprint==65.1
webencodings==0.5.1
xattr==1.1.4
xhtml2pdf>=0.2.13
zipp==3.22.0
zopfli==0.2.3.post1
zstandard==0.23.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
django-filter>=23.5
django-cors-headers>=4.3.1
celery>=5.3.6
redis>=5.0.1
django-celery-results>=2.5.1
django-celery-beat>=2.5.0
